% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_help.R
\name{compute_distribution_teststat_fixed_es}
\alias{compute_distribution_teststat_fixed_es}
\title{Compute the mean and sd of the score test statistic with fixed effect size}
\usage{
compute_distribution_teststat_fixed_es(
  fold_change,
  expression_mean,
  expression_size,
  num_trt_cells,
  num_cntrl_cells,
  num_cells
)
}
\arguments{
\item{fold_change}{A vector consisting of the fold changes for each gRNA-gene pair}

\item{expression_mean}{Mean gene expression}

\item{expression_size}{Size parameter in NB distribution}

\item{num_trt_cells}{Number of treatment cells in score test}

\item{num_cntrl_cells}{Number of control cells in score test}

\item{num_cells}{A vector consisting of the number of cells for each guide RNA}
}
\value{
A list including mean and sd of the test statistic
}
\description{
Compute the mean and sd of the score test statistic with fixed effect size
}
