% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_help.R
\name{compute_distribution_teststat_fixed_es_random_assignment}
\alias{compute_distribution_teststat_fixed_es_random_assignment}
\title{Compute the mean and sd of the score test statistic with fixed effect size and random gRNA assignment}
\usage{
compute_distribution_teststat_fixed_es_random_assignment(
  fold_change,
  expression_mean,
  expression_size,
  num_cntrl_cells,
  mean_num_cells,
  sd_num_cells,
  B = 1000
)
}
\arguments{
\item{fold_change}{A vector consisting of the fold changes for each gRNA-gene pair}

\item{expression_mean}{Mean gene expression}

\item{expression_size}{Size parameter in NB distribution}

\item{num_cntrl_cells}{Number of control cells in score test}

\item{mean_num_cells}{Mean of number of treatment cells}

\item{sd_num_cells}{Standard deviation of number of treatment cells}

\item{B}{Number of Monte Carlo replications to estiamte variance}
}
\value{
A list including mean and sd of the test statistic
}
\description{
Compute the mean and sd of the score test statistic with fixed effect size and random gRNA assignment
}
