% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_monte_carlo_teststat_cpp}
\alias{compute_monte_carlo_teststat_cpp}
\title{Compute Monte Carlo test statistics for power analysis with random effect sizes}
\usage{
compute_monte_carlo_teststat_cpp(
  fc_expression_df,
  library_size,
  num_trt_cells,
  num_cntrl_cells
)
}
\arguments{
\item{fc_expression_df}{Data frame with fold change and expression information}

\item{library_size}{Library size parameter}

\item{num_trt_cells}{Number of treatment cells}

\item{num_cntrl_cells}{Number of control cells}
}
\value{
List with Monte Carlo mean and standard deviation vectors
}
\description{
Compute Monte Carlo test statistics for power analysis with random effect sizes
}
