% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scratch.R
\name{compute_power_grid}
\alias{compute_power_grid}
\title{Compute power in underspecified setup}
\usage{
compute_power_grid(
  cells_reads_df,
  num_targets = 100,
  gRNAs_per_target = 4,
  non_targeting_gRNAs = 10,
  num_pairs = 1000,
  tpm_threshold = 10,
  fdr_target = 0.05,
  fc_mean = 0.85,
  fc_sd = 0.15,
  prop_non_null = 0.1,
  MOI = 10,
  biological_system = "K562",
  experimental_platform = "10x Chromium v3",
  side = "left",
  control_group = "complement",
  B = 500
)
}
\arguments{
\item{cells_reads_df}{A dataframe including columns num_total_cells and reads_per_cell}

\item{num_targets}{Number of targets}

\item{gRNAs_per_target}{Number of gRNAs per target}

\item{non_targeting_gRNAs}{Number of non-targeting gRNAs}

\item{num_pairs}{Number of pairs analyzed}

\item{tpm_threshold}{Minimum TPM threshold}

\item{fdr_target}{FDR target level}

\item{fc_mean}{Fold-change mean}

\item{fc_sd}{Fold-change SD}

\item{prop_non_null}{Proportion of non-null pairs}

\item{MOI}{Multiplicity of infection}

\item{biological_system}{Biological system}

\item{experimental_platform}{Experimental platform}

\item{side}{(Optional) A character string specifying the side of the test, either "left", "right", or "both"; defaults to "both"}

\item{control_group}{A character string specifying the control group, either "complement" or "nt_cells"}

\item{B}{(Optional) A scalar specifying how many Monte Carlo samples used}
}
\value{
A dataframe concatenating the overall_power, power_on_fc and power_on_pi
}
\description{
Compute power in underspecified setup
}
