% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_computation.R
\name{compute_QC}
\alias{compute_QC}
\title{Compute QC probability for each enhancer-gene pair}
\usage{
compute_QC(
  fold_change_mean,
  expression_mean,
  expression_size,
  num_cntrl_cells,
  num_trt_cells,
  n_nonzero_trt_thresh,
  n_nonzero_cntrl_thresh
)
}
\arguments{
\item{fold_change_mean}{A numeric value to use for mean effect size for all element-gene pairs}

\item{expression_mean}{Mean gene expression}

\item{expression_size}{Size parameter in NB distribution}

\item{num_cntrl_cells}{Number of control cells in score test}

\item{num_trt_cells}{Number of treatment cells in score test}

\item{n_nonzero_trt_thresh}{(Optional) An integer specifying the sceptre QC parameter of the same name; defaults to 7}

\item{n_nonzero_cntrl_thresh}{(Optional) An integer specifying the sceptre QC parameter of the same name; defaults to 7}
}
\value{
Probability of a enhancer-gene pair being filtered due to QC
}
\description{
Compute QC probability for each enhancer-gene pair
}
