% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_computation.R
\name{compute_QC_fixed_es}
\alias{compute_QC_fixed_es}
\title{Title}
\usage{
compute_QC_fixed_es(
  fold_change,
  expression_mean,
  expression_size,
  num_cntrl_cells,
  num_cells,
  n_nonzero_trt_thresh = 7,
  n_nonzero_cntrl_thresh = 7
)
}
\arguments{
\item{fold_change}{A vector consisting of the fold changes for each gRNA-gene pair}

\item{expression_mean}{Mean gene expression}

\item{expression_size}{Size parameter in NB distribution}

\item{num_cntrl_cells}{Number of control cells in score test}

\item{num_cells}{A vector consisting of the number of cells for each guide RNA}

\item{n_nonzero_trt_thresh}{(Optional) An integer specifying the sceptre QC parameter of the same name; defaults to 7}

\item{n_nonzero_cntrl_thresh}{(Optional) An integer specifying the sceptre QC parameter of the same name; defaults to 7}
}
\value{
Probability of a enhancer-gene pair being filtered due to QC
}
\description{
Title
}
