% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_computation.R
\name{compute_zero_prob}
\alias{compute_zero_prob}
\title{Compute probability mass of NB distribution at zero}
\usage{
compute_zero_prob(fold_change_mean, expression_mean, expression_size)
}
\arguments{
\item{fold_change_mean}{A numeric value to use for mean effect size for all element-gene pairs}

\item{expression_mean}{Mean gene expression}

\item{expression_size}{Size parameter in NB distribution}
}
\value{
Probability of a NB variable being 0
}
\description{
Compute probability mass of NB distribution at zero
}
