% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_help.R
\name{distribution_teststat}
\alias{distribution_teststat}
\title{Compute the mean and sd approximation of the test statistic.}
\usage{
distribution_teststat(
  control_cell_vec,
  target_cell_vec,
  target_cell_vec_sq,
  baseline_expression,
  size_parameter,
  effect_size_mean,
  effect_size_sd
)
}
\arguments{
\item{control_cell_vec}{Control cell size for L elements (vector; length L). Vector should be named
for grna_target.}

\item{target_cell_vec}{Number of cell per element (vector; length J).}

\item{target_cell_vec_sq}{Squared number of cell per element (vector; length J).}

\item{baseline_expression}{Baseline mean expression in the group of control cells.}

\item{size_parameter}{Size parameter for J genes (vector; length J). Names should be same as \code{relative_expression}.}

\item{effect_size_mean}{Mean fold change for each (element, gene) (matrix; row: L genomic elements;
column: J genes). Rownames and colnames should be specified. The rowname should look like "enh1"
and column name should be ensembl form or like "gene1". Alternatively, a scalar value can just be
specified if all enhancer-gene pairs share the same effect size.}

\item{effect_size_sd}{Sd fold change for each (element, gene). Format should be the same as \code{effect_size_mean}.}
}
\value{
Mean and sd vector of length (L x J).
}
\description{
Compute the mean and sd approximation of the test statistic.
}
