% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_help.R
\name{efficiency_computation}
\alias{efficiency_computation}
\title{Compute the weighted gRNA efficiency by the cell size}
\usage{
efficiency_computation(pi_mat, cell_mat)
}
\arguments{
\item{pi_mat}{gRNA efficiency matrix of dim L by K (row: genomic elements, column: different gRNA)}

\item{cell_mat}{Cell size matrix of dim L by K (row: genomic elements, column: different gRNA)}
}
\value{
A vector of weighted gRNA efficiency vector by cell size (of length L)
}
\description{
Compute the weighted gRNA efficiency by the cell size
}
