% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/James_data.R
\name{generate_James_data}
\alias{generate_James_data}
\title{This is a function generating data from James' DGP}
\usage{
generate_James_data(
  effect_size,
  guide_sd,
  discovery_pairs,
  sce,
  grna_target_data_frame,
  trt_group,
  size_factor = TRUE
)
}
\arguments{
\item{effect_size}{Effect size: a scalar}

\item{guide_sd}{Guide RNA sd}

\item{discovery_pairs}{Discovery pairs}

\item{sce}{sce object}

\item{grna_target_data_frame}{grna target dataframe}

\item{trt_group}{treatment group}

\item{size_factor}{A logic value; if TRUE, use the one in sce; false, set to 1}
}
\value{
A lsit of information
}
\description{
This is a function generating data from James' DGP
}
