% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{input_check}
\alias{input_check}
\title{Input checking function (currently only available for input_check)}
\usage{
input_check(
  perturb_type = NULL,
  target_cell_df = NULL,
  control_cell_vec = NULL,
  size_parameter = NULL,
  baseline_expression = NULL,
  effect_size_mean = NULL,
  effect_size_sd = NULL,
  n_nonzero_trt_thresh = NULL,
  n_nonzero_cntrl_thresh = NULL,
  side = NULL,
  multiple_testing_method = NULL,
  multiple_testing_alpha = NULL,
  cutoff = NULL,
  intermediate_outcome = NULL
)
}
\arguments{
\item{perturb_type}{Either CRISPRi or CRISPRko; currently can only handle CRISPRi.}

\item{target_cell_df}{Perturb cell size for L elements (dataframe). The data frame should include
three columns: grna_id (value should be like "gRNAd" where d is an integer), grna_target (value should
be like "enhl" where l is an integer) and num_cells (value should be integer). Thus each row of the data
frame includes the number of cells for each (gRNA, element) pair.}

\item{control_cell_vec}{Control cell size for L elements (vector; length L). Vector should be named
for grna_target.}

\item{size_parameter}{Size parameter for J genes (vector; length J). Names should be same as \code{relative_expression}.}

\item{baseline_expression}{Baseline mean expression in the group of control cells.}

\item{effect_size_mean}{Mean fold change for each (element, gene) (matrix; row: L genomic elements;
column: J genes). Rownames and colnames should be specified. The rowname should look like "enh1"
and column name should be ensembl form or like "gene1". Alternatively, a scalar value can just be
specified if all enhancer-gene pairs share the same effect size.}

\item{effect_size_sd}{Sd fold change for each (element, gene). Format should be the same as \code{effect_size_mean}.}

\item{n_nonzero_trt_thresh}{QC threshold for treatment cell.}

\item{n_nonzero_cntrl_thresh}{QC threshold for control cell.}

\item{side}{Left, right or both.}

\item{multiple_testing_method}{Multiplicity correction; default is BH.}

\item{multiple_testing_alpha}{Significance level of interest (FDR of interest).}

\item{cutoff}{Cutoff for p-values to reject the test.}

\item{intermediate_outcome}{A logic value indicating if only mean, sd of test statistics and QC probability are desired.}
}
\description{
Input checking function (currently only available for input_check)
}
