% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{input_check_library_estimation}
\alias{input_check_library_estimation}
\title{Input checking function for library_estimation}
\usage{
input_check_library_estimation(QC_data = NULL)
}
\arguments{
\item{QC_data}{Data frame. UMI-level molecule information from
\code{\link{obtain_qc_read_umi_table}} containing columns \code{num_reads},
\code{UMI_id}, \code{cell_id}, and \code{response_id}.}
}
\description{
Input checking function for library_estimation
}
\keyword{internal}
