% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{input_check_posthoc}
\alias{input_check_posthoc}
\title{Input checking function for compute_power_posthoc}
\usage{
input_check_posthoc(
  discovery_pairs = NULL,
  cells_per_grna = NULL,
  baseline_expression_stats = NULL,
  control_group = NULL,
  fold_change_mean = NULL,
  fold_change_sd = NULL,
  num_total_cells = NULL,
  cutoff = NULL,
  n_nonzero_trt_thresh = NULL,
  n_nonzero_cntrl_thresh = NULL,
  side = NULL,
  multiple_testing_method = NULL,
  multiple_testing_alpha = NULL
)
}
\arguments{
\item{discovery_pairs}{A data frame specifying which element-gene pairs to consider, with columns \code{grna_target} and \code{response_id}}

\item{cells_per_grna}{A data frame specifying how many cells contain each gRNA, with columns \code{grna_id}, \code{grna_target}, and \code{num_cells}}

\item{baseline_expression_stats}{A data frame specifying the baseline expression statistics for each gene, with columns \code{response_id}, \code{expression_mean}, and \code{expression_size}}

\item{control_group}{A character string specifying the control group, either "complement" or "nt_cells"}

\item{fold_change_mean}{A numeric value to use for mean effect size for all element-gene pairs}

\item{fold_change_sd}{A numeric value to use for standard deviation of effect size for all element-gene pairs}

\item{num_total_cells}{(Required only if control_group == "complement") A positive integer specifying the total number of cells in the experiment}

\item{cutoff}{(Optional) A numeric value between 0 and 1 to use as the p-value cutoff}

\item{n_nonzero_trt_thresh}{(Optional) An integer specifying the sceptre QC parameter of the same name; defaults to 7}

\item{n_nonzero_cntrl_thresh}{(Optional) An integer specifying the sceptre QC parameter of the same name; defaults to 7}

\item{side}{(Optional) A character string specifying the side of the test, either "left", "right", or "both"; defaults to "both"}

\item{multiple_testing_method}{(Optional) A character string specifying the multiple testing correction method to use, either "BH" or "bonferroni"; defaults to "BH"}

\item{multiple_testing_alpha}{(Optional) A numeric value between 0 and 1 specifying the alpha level for multiple testing correction; defaults to 0.1}
}
\description{
Input checking function for compute_power_posthoc
}
\keyword{internal}
