% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimation_help.R
\name{obtain_qc_data}
\alias{obtain_qc_data}
\title{Obtain a data frame with information of all QC'd reads.}
\usage{
obtain_qc_data(path_to_outs_folder)
}
\arguments{
\item{path_to_outs_folder}{The path to the outs folder of the cellranger output.}
}
\value{
A data frame with columns \code{num_reads}, \code{UMI_id}, \code{cell_id}, and \code{response_id}
}
\description{
Obtain a data frame with information of all QC'd reads.
}
