% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_data_preprocessing.R
\name{obtain_random_pairs}
\alias{obtain_random_pairs}
\title{Generate random gRNA–gene discovery pairs for control or simulation.}
\usage{
obtain_random_pairs(num_targets, pairs_per_target, gene_info)
}
\arguments{
\item{num_targets}{Integer. Number of pseudo gRNAs to simulate.}

\item{pairs_per_target}{Integer. Number of genes to assign per gRNA.}

\item{gene_info}{Data frame. Must contain a column named \code{response_id} with gene names.}
}
\value{
A data frame with columns \code{response_id} and \code{grna_id}, each row representing a pseudo discovery pair.
}
\description{
Generate random gRNA–gene discovery pairs for control or simulation.
}
