% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power_function}
\alias{power_function}
\title{Power function for CRISPR screen experimental design.}
\usage{
power_function(
  target_cell_df,
  control_cell_vec,
  library_size = NULL,
  relative_expression,
  size_parameter,
  effect_size_mean = 0.8,
  effect_size_sd = 0,
  n_nonzero_trt_thresh = 7,
  n_nonzero_cntrl_thresh = 7,
  UMI_s = NULL,
  recovery_rate = NULL,
  doublet_rate = NULL,
  doublet_factor = NULL,
  planned_read = NULL,
  mapping_efficiency = NULL,
  side,
  multiple_testing_method = "BH",
  multiple_testing_alpha = 0.1,
  intermediate_outcome = FALSE
)
}
\arguments{
\item{target_cell_df}{Perturb cell size for L elements (dataframe). The data frame should include
three columns: grna_id (value should be like "gRNAd" where d is an integer), grna_target (value should
be like "enhl" where l is an integer) and num_cells (value should be integer). Thus each row of the data
frame includes the number of cells for each (gRNA, element) pair.}

\item{control_cell_vec}{Control cell size for L elements (vector; length L). Vector should be named
for grna_target.}

\item{library_size}{Averaged library size (scalar; positive valued).}

\item{relative_expression}{Baseline relative expression level for J genes (named vector; of length J).
Names of the vector should be different genes and the format should be consistent with the column names of
\code{effect_size_mean} and \code{effect_size_sd}.}

\item{size_parameter}{Size parameter for J genes (vector; length J). Names should be same as \code{relative_expression}.}

\item{effect_size_mean}{Mean fold change for each (element, gene) (matrix; row: L genomic elements;
column: J genes). Rownames and colnames should be specified. The rowname should look like "enh1"
and column name should be ensembl form or like "gene1". Alternatively, a scalar value can just be
specified if all enhancer-gene pairs share the same effect size.}

\item{effect_size_sd}{Sd fold change for each (element, gene). Format should be the same as \code{effect_size_mean}.}

\item{n_nonzero_trt_thresh}{QC threshold for treatment cell.}

\item{n_nonzero_cntrl_thresh}{QC threshold for control cell.}

\item{UMI_s}{Estimate for UMI count per singlet for the underlying type of cell.}

\item{recovery_rate}{Recovery rate for cells surviving the library preparation.}

\item{doublet_rate}{Doublet rate for droplet containing more than one cell.}

\item{doublet_factor}{Ratio of averaged UMI count per doublet to UMI per singlet.}

\item{planned_read}{Planned total sequencing reads.}

\item{mapping_efficiency}{Mapping efficiency for sequenced reads.}

\item{side}{Left, right or both.}

\item{multiple_testing_method}{Multiplicity correction; default is BH.}

\item{multiple_testing_alpha}{Significance level of interest (FDR of interest).}

\item{intermediate_outcome}{A logic value indicating if only mean, sd of test statistics and QC probability are desired.}
}
\value{
The output format is the same as \code{\link[=compute_power]{compute_power()}}.
}
\description{
Power function for CRISPR screen experimental design.
}
