% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_help.R
\name{read_per_cell}
\alias{read_per_cell}
\title{Compute averaged number of sequencing reads for each cell after mapping}
\usage{
read_per_cell(planned_read, mapping_efficiency, planned_cell, recovery_rate)
}
\arguments{
\item{planned_read}{Number of reads planned before the experiment}

\item{mapping_efficiency}{Mapping efficiency for sequenced reads}

\item{planned_cell}{Number of cells planned before the experiment}

\item{recovery_rate}{Fraction of cells surviving after the library preparation}
}
\value{
Averaged number of sequencing reads for each cell
}
\description{
Compute averaged number of sequencing reads for each cell after mapping
}
