% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_help.R
\name{rejection_computation}
\alias{rejection_computation}
\title{Compute the rejection probability.}
\usage{
rejection_computation(mean_list, sd_list, side, cutoff)
}
\arguments{
\item{mean_list}{Asymptotic mean of test statistic (vector; length L x J).}

\item{sd_list}{Asymptotic sd of test statistic (vector; length L x J).}

\item{side}{Left, right or both.}

\item{cutoff}{Cutoff for p-values to reject the test.}
}
\value{
The rejection probablity.
}
\description{
Compute the rejection probability.
}
