% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/James_DGP.R
\name{sim_tapseq_sce}
\alias{sim_tapseq_sce}
\title{Simulate Perturb-seq data based on a SingleCellExperiment object with added mean expression,
dispersion and size factors from original real data}
\usage{
sim_tapseq_sce(sce, effect_size_mat)
}
\arguments{
\item{sce}{Perturb-seq SingleCellExperiment object}

\item{effect_size_mat}{Effect size matrix}
}
\description{
Simulate Perturb-seq data based on a SingleCellExperiment object with added mean expression,
dispersion and size factors from original real data
}
