% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/James_DGP.R
\name{simulate_tapseq_counts}
\alias{simulate_tapseq_counts}
\title{Simulate Perturb-seq data}
\usage{
simulate_tapseq_counts(
  gene_means,
  gene_dispersions,
  cell_size_factors,
  effect_size_mat,
  gene_ids = names(gene_means),
  cell_ids = names(cell_size_factors)
)
}
\value{
A matrix with simulated Perturb-seq UMI counts
}
\description{
Simulate Perturb-seq UMI counts for a given number of perturbed cells with a specified effect size on
specified genes.
}
