% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimation_help.R
\name{summary_data}
\alias{summary_data}
\title{Obtain the summary statistics of the QC'd data.}
\usage{
summary_data(QC_data)
}
\arguments{
\item{QC_data}{The QC'd data coming from the function obtain_qc_data}
}
\value{
A named vector with the number of cells and average reads per cell.
}
\description{
Obtain the summary statistics of the QC'd data.
}
