% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc_help.R
\name{var_nb}
\alias{var_nb}
\title{Variance of NB distribution}
\usage{
var_nb(mean, size)
}
\arguments{
\item{mean}{mean gene expression.}

\item{size}{size parameter.}
}
\value{
variance of NB distribution.
}
\description{
Variance of NB distribution
}
\keyword{internal}
