% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc_help.R
\name{adjusted_cutoff}
\alias{adjusted_cutoff}
\title{Compute the adjusted significance level with either BH or Bonferroni procedure.}
\usage{
adjusted_cutoff(
  mean_list,
  sd_list,
  multiple_testing_alpha,
  multiple_testing_method,
  side,
  QC_prob
)
}
\arguments{
\item{mean_list}{Asymptotic mean of test statistic}

\item{sd_list}{Asymptotic sd of test statistic}

\item{multiple_testing_alpha}{(Optional) A numeric value between 0 and 1 specifying the alpha level for multiple testing correction; defaults to 0.1}

\item{multiple_testing_method}{(Optional) A character string specifying the multiple testing correction method to use, either "BH" or "bonferroni"; defaults to "BH"}

\item{side}{(Optional) A character string specifying the side of the test, either "left", "right", or "both"; defaults to "both"}

\item{QC_prob}{The probability of failing QC}
}
\value{
The adjusted significance level.
}
\description{
Compute the adjusted significance level with either BH or Bonferroni procedure.
}
