% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_help.R
\name{adjusted_cutoff}
\alias{adjusted_cutoff}
\title{Compute the adjusted significance level with either BH or Bonferroni procedure}
\usage{
adjusted_cutoff(mean_list, sig_level, correction, sideness, QC_prob)
}
\arguments{
\item{mean_list}{List of mean under local alternative}

\item{sig_level}{Significance level imposed by users}

\item{correction}{Either BH or Bonferroni}

\item{sideness}{Sideness of the testing procedure}

\item{QC_prob}{QC probability}
}
\value{
The adjusted significance level
}
\description{
Compute the adjusted significance level with either BH or Bonferroni procedure
}
