% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_help.R
\name{adjusted_power}
\alias{adjusted_power}
\title{Adjusted power based on adjusted significance level.}
\usage{
adjusted_power(
  mean_list,
  sd_list,
  multiple_testing_alpha = NULL,
  multiple_testing_method = NULL,
  side,
  QC_prob,
  cutoff = NULL
)
}
\arguments{
\item{mean_list}{Asymptotic mean of test statistic (vector; length L x J).}

\item{sd_list}{Asymptotic sd of test statistic (vector; length L x J).}

\item{multiple_testing_alpha}{Significance level of interest (FDR of interest).}

\item{multiple_testing_method}{Multiplicity correction; default is BH.}

\item{side}{Left, right or both.}

\item{QC_prob}{QC probability for each enhancer-gene pair (vector; length L x J).}

\item{cutoff}{Cutoff for p-values to reject the test.}
}
\value{
Adjusted power list including adjusted power and discovery size estimate.
}
\description{
Adjusted power based on adjusted significance level.
}
