% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_help.R
\name{adjusted_power}
\alias{adjusted_power}
\title{Adjusted power based on adjusted significance level}
\usage{
adjusted_power(mean_list, sig_level, correction, sideness, QC_prob)
}
\arguments{
\item{mean_list}{List of mean under local alternative}

\item{sig_level}{Significance level imposed by users}

\item{correction}{Either BH or Bonferroni}

\item{sideness}{Sideness of the testing procedure}

\item{QC_prob}{QC probability}
}
\value{
Adjusted power list including adjusted power and discovery size estimate
}
\description{
Adjusted power based on adjusted significance level
}
