% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_help.R
\name{BH_cutoff}
\alias{BH_cutoff}
\title{Compute the adjusted cutoff/significance level applying BH procedure.}
\usage{
BH_cutoff(mean_list, sd_list, side, multiple_testing_alpha, QC_prob)
}
\arguments{
\item{mean_list}{Asymptotic mean of test statistic}

\item{sd_list}{Asymptotic sd of test statistic}

\item{side}{(Optional) A character string specifying the side of the test, either "left", "right", or "both"; defaults to "both"}

\item{multiple_testing_alpha}{(Optional) A numeric value between 0 and 1 specifying the alpha level for multiple testing correction; defaults to 0.1}
}
\value{
Adjusted cutoff/significance level.
}
\description{
Compute the adjusted cutoff/significance level applying BH procedure.
}
