% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_help.R
\name{BH_cutoff}
\alias{BH_cutoff}
\title{Compute the adjusted cutoff/significance level applying BH procedure}
\usage{
BH_cutoff(mean_list, sideness, sig_level)
}
\arguments{
\item{mean_list}{List of mean under local alternative}

\item{sideness}{Sideness of the testing procedure}

\item{sig_level}{Significance level imposed by users}
}
\value{
Adjusted cutoff/significance level
}
\description{
Compute the adjusted cutoff/significance level applying BH procedure
}
