% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc_help.R
\name{BH_cutoff_bisection}
\alias{BH_cutoff_bisection}
\title{Benjamini–Hochberg cutoff with bisection search (C++ back-end)}
\usage{
BH_cutoff_bisection(mean_list, sd_list, side, multiple_testing_alpha, QC_prob)
}
\arguments{
\item{mean_list}{Asymptotic mean of test statistic}

\item{sd_list}{Asymptotic sd of test statistic}

\item{side}{(Optional) A character string specifying the side of the test, either "left", "right", or "both"; defaults to "both"}

\item{multiple_testing_alpha}{(Optional) A numeric value between 0 and 1 specifying the alpha level for multiple testing correction; defaults to 0.1}

\item{QC_prob}{The probability of failing QC}
}
\value{
Adjusted cutoff/significance level.
}
\description{
Thin wrapper that validates inputs and forwards to the compiled routine.
}
