% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc_help.R
\name{compute_distribution_teststat}
\alias{compute_distribution_teststat}
\title{Compute mean and sd of the score test statistic}
\usage{
compute_distribution_teststat(
  num_trt_cells,
  num_cntrl_cells,
  num_trt_cells_sq,
  expression_mean,
  expression_size,
  fold_change_mean,
  fold_change_sd
)
}
\arguments{
\item{num_trt_cells}{Number of treatment cells in score test}

\item{num_cntrl_cells}{Number of control cells in score test}

\item{num_trt_cells_sq}{Squared number of control cells in score test}

\item{expression_mean}{Mean gene expression}

\item{expression_size}{Size parameter in NB distribution}

\item{fold_change_mean}{A numeric value to use for mean effect size for all element-gene pairs}

\item{fold_change_sd}{A numeric value to use for standard deviation of effect size for all element-gene pairs}
}
\value{
A list including mean and sd of the test statistic
}
\description{
Compute mean and sd of the score test statistic
}
