% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{compute_power}
\alias{compute_power}
\title{Computing power for each (element, gene) pair}
\usage{
compute_power(
  perturb_type,
  control_cell_vec,
  target_cell_mat,
  library_size,
  relative_expression,
  size_parameter,
  effect_size_mean,
  effect_size_sd,
  QC,
  n_nonzero_trt,
  n_nonzero_ctl,
  sideness,
  correction = NULL,
  sig_level = NULL,
  cutoff = NULL,
  intermediate_outcome = FALSE
)
}
\arguments{
\item{perturb_type}{either CRISPRi or CRISPRko; currently can only handle CRISPRi}

\item{control_cell_vec}{Control cell size for L elements (vector; length L)}

\item{target_cell_mat}{Perturb cell size for L elements (matrix; row: L genomic elements; column: K gRNA libraries)}

\item{library_size}{Averaged library size (scalar; positive valued)}

\item{relative_expression}{Baseline relative expression level for J genes (scalar; valued between 0 and 1)}

\item{size_parameter}{Size parameter for J genes (vector; length J)}

\item{effect_size_mean}{Mean fold change for each (element, gene) (matrix; row: L genomic elements; column: J genes)}

\item{effect_size_sd}{Sd fold change for each (element, gene) (matrix; row: L genomic elements; column: J genes))}

\item{QC}{A logic value; if TRUE, then do quality control}

\item{n_nonzero_trt}{QC threshold for treatment cell}

\item{n_nonzero_ctl}{QC threshold for control cell}

\item{sideness}{Left, right or both}

\item{correction}{Multiplicity correction; default is BH}

\item{sig_level}{Significance leve of interest (FDR of interest)}

\item{cutoff}{Cutoff for p-values to reject the test}

\item{intermediate_outcome}{A logic value indicating if only mean, sd of test statistics and QC are desired}
}
\value{
A dataframe or a list epending on intermediate_outcome.
}
\description{
Computing power for each (element, gene) pair
}
