% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{compute_power}
\alias{compute_power}
\title{Computing power for each (enhancer, gene) pair.}
\usage{
compute_power(
  control_cell_vec,
  target_cell_df,
  baseline_expression,
  size_parameter,
  effect_size_mean,
  effect_size_sd,
  n_nonzero_trt_thresh = 7,
  n_nonzero_cntrl_thresh = 7,
  side = "both",
  multiple_testing_method = "BH",
  multiple_testing_alpha = 0.1,
  cutoff = NULL,
  intermediate_outcome = FALSE
)
}
\arguments{
\item{control_cell_vec}{Control cell size for L elements (vector; length L). Vector should be named
for grna_target.}

\item{target_cell_df}{Perturb cell size for L elements (dataframe). The data frame should include
three columns: grna_id (value should be like "gRNAd" where d is an integer), grna_target (value should
be like "enhl" where l is an integer) and num_cells (value should be integer). Thus each row of the data
frame includes the number of cells for each (gRNA, element) pair.}

\item{baseline_expression}{Baseline mean expression in the group of control cells.}

\item{size_parameter}{Size parameter for J genes (vector; length J). Names should be same as \code{relative_expression}.}

\item{effect_size_mean}{Mean fold change for each (element, gene) (matrix; row: L genomic elements;
column: J genes). Rownames and colnames should be specified. The rowname should look like "enh1"
and column name should be ensembl form or like "gene1". Alternatively, a scalar value can just be
specified if all enhancer-gene pairs share the same effect size.}

\item{effect_size_sd}{Sd fold change for each (element, gene). Format should be the same as \code{effect_size_mean}.}

\item{n_nonzero_trt_thresh}{QC threshold for treatment cell.}

\item{n_nonzero_cntrl_thresh}{QC threshold for control cell.}

\item{side}{Left, right or both.}

\item{multiple_testing_method}{Multiplicity correction; default is BH.}

\item{multiple_testing_alpha}{Significance level of interest (FDR of interest).}

\item{cutoff}{Cutoff for p-values to reject the test.}

\item{intermediate_outcome}{A logic value indicating if only mean, sd of test statistics and QC probability are desired.}
}
\value{
A dataframe or a list depending on \code{intermediate_outcome}:
}
\description{
This function have two functionalities: (1, intermediate goal) compute the mean
and sd of the test statistic; (2, final goal) output the power for each enhancer-gene
pair.
}
\section{Dataframe}{
 if \code{intermediate_outcome} is TRUE, only mean, sd of the score tests and QC probability will be outputted.
}

\section{List}{
 otherwise, the final power for each element-gene pair will be outputted, together with the estimated discovery set.
}

