% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_help.R
\name{distribution_teststat}
\alias{distribution_teststat}
\title{Compute the mean and sd approximation of the test statistic}
\usage{
distribution_teststat(
  control_cell_vec,
  target_cell_mat,
  size_parameter,
  effect_size_mean,
  effect_size_sd,
  pooled_mean,
  trt_mean,
  ctl_mean
)
}
\arguments{
\item{control_cell_vec}{Vector including number of cell negative control group for each element}

\item{target_cell_mat}{Matrix including number of cell in group (l,k) (row: L genomic elements; column: K gRNA libraries)}

\item{effect_size_sd}{Sd matrix for random effect size in treatment group (L by J dimension)}

\item{pooled_mean}{Pooled mean expression of treatment and control (vector; length J)}

\item{trt_mean}{Treatment mean expression (vector; length J)}

\item{ctl_mean}{Control mean expression (vector; length J)}
}
\value{
Mean and sd vector of length (L times J)
}
\description{
Compute the mean and sd approximation of the test statistic
}
