% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_help.R
\name{FDP_estimate}
\alias{FDP_estimate}
\title{FDP estimate based on rejection probability.}
\usage{
FDP_estimate(mean_list, sd_list, side, cutoff, QC_prob)
}
\arguments{
\item{mean_list}{Asymptotic mean of test statistic (vector; length L x J).}

\item{sd_list}{Asymptotic sd of test statistic (vector; length L x J).}

\item{side}{Left, right or both.}

\item{cutoff}{Cutoff for p-values to reject the test.}

\item{QC_prob}{QC probability for each enhancer-gene pair (vector; length L x J).}
}
\value{
FDP estimate.
}
\description{
FDP estimate based on rejection probability.
}
