% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_help.R
\name{FDP_estimate}
\alias{FDP_estimate}
\title{FDP estimate based on rejection probability}
\usage{
FDP_estimate(mean_list, sd_list, sideness, sig_level, QC_prob)
}
\arguments{
\item{mean_list}{List of mean under local alternative}

\item{sd_list}{List of sd under local alternative}

\item{sideness}{Sideness of the testing procedure}

\item{sig_level}{Significance level imposed by users}

\item{QC_prob}{QC probability}
}
\value{
FDP estimate
}
\description{
FDP estimate based on rejection probability
}
