% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_help.R
\name{gene_part_computation}
\alias{gene_part_computation}
\title{Compute the gene expression related quantity in the power function}
\usage{
gene_part_computation(
  expression_level_list,
  library_size,
  effect_size,
  num_control,
  num_trt,
  size_factor_trt,
  size_factor_ctl
)
}
\arguments{
\item{expression_level_list}{List of relative gene expression level}

\item{library_size}{Library size for each cell considered}

\item{effect_size}{Effect size matrix (L by J)}

\item{num_control}{Control cell vector (of length L)}

\item{num_trt}{Treatment cell vector (of length L)}

\item{size_factor_trt}{size factor matrix for treatment group}

\item{size_facotr_ctl}{size factor matrix for control group}
}
\value{
The gene expression related part in the power function
}
\description{
Compute the gene expression related quantity in the power function
}
