% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_help.R
\name{gene_part_computation}
\alias{gene_part_computation}
\title{Compute the gene expression related quantity in the power function}
\usage{
gene_part_computation(expression_level_list, dispersion_list, library_size)
}
\arguments{
\item{expression_level_list}{List of relative gene expression level}

\item{dispersion_list}{Dispersion parameter for each gene}

\item{library_size}{Library size for each cell considered}
}
\value{
The gene expression related part in the power function
}
\description{
Compute the gene expression related quantity in the power function
}
