% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimation.R
\name{library_estimation}
\alias{library_estimation}
\title{Compute the average total UMI per cell and UMI variation parameters.}
\usage{
library_estimation(QC_data, downsample_ratio = 0.7, D2_rough = 0.3)
}
\arguments{
\item{QC_data}{The QC'd data coming from the function obtain_qc_data.}

\item{downsample_ratio}{The ratio of the size of the downsampled dataset to the one of the original dataset.}

\item{D2_rough}{The rough estimate of D2 in the S-M curve model.}
}
\value{
Named vector of average total UMI per cell and UMI variation.
}
\description{
Compute the average total UMI per cell and UMI variation parameters.
}
