% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimation.R
\name{library_estimation}
\alias{library_estimation}
\title{Compute the average total UMI per cell and UMI variation parameters.}
\usage{
library_estimation(QC_data, library_model)
}
\arguments{
\item{QC_data}{QC'd data coming from the function obtain_qc_data in parameter_estimation_help.R}

\item{library_model}{S-M curve model fitted by the function model_fit in helper-S-M-curve.R}
}
\value{
Named vector of average total UMI per cell and UMI variation.
}
\description{
Compute the average total UMI per cell and UMI variation parameters.
}
