% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_help.R
\name{mean_expression_computation}
\alias{mean_expression_computation}
\title{Compute the gene expression related quantity in the power function.}
\usage{
mean_expression_computation(
  baseline_expression,
  effect_size_mean,
  num_control,
  num_trt
)
}
\arguments{
\item{baseline_expression}{Baseline gene expression in the group of control cells.}

\item{effect_size_mean}{Mean fold change for each (element, gene) (matrix; row: L genomic elements; column: J genes).}

\item{num_control}{Number of control cell (vector; length L).}

\item{num_trt}{Number of perturb cell (vector; length L).}
}
\value{
The gene expression related part in the power function.
}
\description{
Compute the gene expression related quantity in the power function.
}
