% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimation_help.R
\name{obtain_expression_information}
\alias{obtain_expression_information}
\title{Obtain gene-level expression and filtering information based on TPM.}
\usage{
obtain_expression_information(
  TPM_thres = 10,
  response_matrix = NULL,
  file_type = "mtx",
  path_to_gene_expression,
  cell_covariates = NULL
)
}
\arguments{
\item{TPM_thres}{TPM threshold for gene filtering (default = 10).}

\item{response_matrix}{Optional expression matrix (genes × cells). If NULL, it will be read from file.}

\item{file_type}{Either "mtx" (default) or "odm", indicating the file format.}

\item{path_to_gene_expression}{Path to the folder containing expression matrix files.}

\item{cell_covariates}{A data frame of cell-level covariates (used in precomputation).}
}
\value{
A data frame with genes passing the TPM threshold, with relative expression and size parameter.
}
\description{
Obtain gene-level expression and filtering information based on TPM.
}
