% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_data_preprocessing.R
\name{obtain_expression_information}
\alias{obtain_expression_information}
\title{Obtain gene-level expression and filtering information based on TPM.}
\usage{
obtain_expression_information(
  TPM_thres = 10,
  response_matrix,
  cell_covariates = NULL
)
}
\arguments{
\item{TPM_thres}{TPM threshold for gene filtering (default = 10).}

\item{response_matrix}{Expression matrix (genes × cells).}

\item{cell_covariates}{Optional cell-level covariates. If NULL, uses log-library size.}
}
\value{
A data frame of genes with relative expression and estimated size (theta).
}
\description{
Obtain gene-level expression and filtering information based on TPM.
}
