% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_data_help.R
\name{obtain_mapping_efficiency}
\alias{obtain_mapping_efficiency}
\title{Mapping efficiency of a Cell Ranger run}
\usage{
obtain_mapping_efficiency(QC_data, path_to_cellranger_output)
}
\arguments{
\item{QC_data}{Output of \code{obtain_qc_read_umi_table()}.}

\item{path_to_cellranger_output}{Folder containing \code{outs/metrics_summary.csv}.}
}
\value{
Numeric proportion: mapped / total reads.
}
\description{
Mapping efficiency of a Cell Ranger run
}
\examples{
# Get mapping efficiency from Cell Ranger output
cellranger_path <- system.file("extdata/cellranger_tiny", package = "perturbplan")
qc_data <- obtain_qc_read_umi_table(cellranger_path)
}
