% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_data_help.R
\name{obtain_qc_read_umi_table}
\alias{obtain_qc_read_umi_table}
\title{Extract QC-filtered molecule information from Cell Ranger HDF5 files}
\usage{
obtain_qc_read_umi_table(path_to_cellranger_output)
}
\arguments{
\item{path_to_cellranger_output}{Character. Folder containing
\code{outs/molecule_info.h5} and \code{outs/filtered_feature_bc_matrix.h5}.}
}
\value{
Data frame with columns \code{num_reads}, \code{UMI_id},
\code{cell_id}, \code{response_id}.
}
\description{
Extract QC-filtered molecule information from Cell Ranger HDF5 files
}
\examples{
# Extract read/UMI information from Cell Ranger output
cellranger_path <- system.file("extdata/cellranger_tiny", package = "perturbplan")
qc_table <- obtain_qc_read_umi_table(cellranger_path)
# Examine the data
head(qc_table)
dim(qc_table)
summary(qc_table)

}
\keyword{internal}
