% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_data_preprocessing.R
\name{obtain_random_pairs}
\alias{obtain_random_pairs}
\title{Simulate random gRNA–gene discovery pairs}
\usage{
obtain_random_pairs(num_targets, pairs_per_target, gene_info)
}
\arguments{
\item{num_targets}{Integer. Number of mock gRNAs.}

\item{pairs_per_target}{Integer. Genes assigned to each gRNA.}

\item{gene_info}{Data frame from \code{obtain_expression_information()}, must
contain \code{response_id}.}
}
\value{
A two-column \code{data.frame} (\code{response_id}, \code{grna_id}).
}
\description{
Simulate random gRNA–gene discovery pairs
}
