% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power_function}
\alias{power_function}
\title{Power function for CRISPR screen experimental design}
\usage{
power_function(
  control_cell_vec,
  target_cell_mat,
  UMI_s,
  library_size = NULL,
  expression_level,
  size_parameter_list,
  gRNA_efficiency_mat,
  recovery_rate,
  doublet_rate,
  doublet_factor,
  planned_read,
  mapping_efficiency,
  effect_size,
  sideness,
  correction = "BH",
  sig_level = 0.1,
  QC = TRUE,
  n_thresh = 7,
  intermediate_outcome = FALSE,
  effect_size_sd = NULL,
  size_factor_list = list(size_factor = matrix(1, nrow = nrow(target_cell_mat), ncol = 2,
    dimnames = list(element = NULL, group = c("trt", "ctl"))), size_factor_sq = matrix(1,
    nrow = nrow(target_cell_mat), ncol = 2, dimnames = list(element = NULL, group =
    c("trt", "ctl")))),
  perturb_type = "CRISPRi"
)
}
\arguments{
\item{control_cell_vec}{Vector including number of cell negative control group for each element}

\item{target_cell_mat}{Matrix including number of cell in group (l,k) (row: L genomic elements; column: K gRNA libraries)}

\item{UMI_s}{Estimate for UMI count per singlet for the underlying type of cell}

\item{library_size}{Library size from pilot data or set as a parameter in simulation study}

\item{expression_level}{Expression level for J genes (values between 0 and 1)}

\item{size_parameter_list}{Size parameter list including for all J genes}

\item{gRNA_efficiency_mat}{gRNA efficiency matrix of dimension L by K}

\item{recovery_rate}{Recovery rate for cells surviving the library preparation}

\item{doublet_rate}{Doublet rate for droplet containing more than one cell}

\item{doublet_factor}{atio of averaged UMI count per doublet to UMI per singlet}

\item{planned_read}{Planned total sequencing reads}

\item{mapping_efficiency}{Mapping efficiency for sequenced reads}

\item{effect_size}{Effect size matrix for each element l and gene j (L by J dimension)}

\item{sideness}{Left, right or both}

\item{correction}{Multiplicity correction, example including BH, bonferroni}

\item{sig_level}{False discovery rate level}

\item{QC}{A logic value; if TRUE, then do quality control}

\item{n_thresh}{QC threshold}

\item{intermediate_outcome}{A logic value indicating if only local mean and QC prob is output}

\item{effect_size_sd}{Sd matrix for random effect size in treatment group (L by J dimension)}

\item{size_factor_list}{List of size factor with two matrices}

\item{perturb_type}{either CRISPRi or CRISPRko; default is CRISPRi}
}
\value{
Either power list or power list and discovery size
}
\description{
Power function for CRISPR screen experimental design
}
