% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power_function}
\alias{power_function}
\title{Power function for CRISPR screen experimental design.}
\usage{
power_function(
  perturb_type = "CRISPRi",
  control_cell_vec,
  target_cell_mat,
  library_size = NULL,
  relative_expression,
  size_parameter,
  effect_size_mean,
  effect_size_sd = matrix(0, nrow = nrow(effect_size_mean), ncol =
    ncol(effect_size_mean)),
  QC = TRUE,
  n_nonzero_trt = 7,
  n_nonzero_ctl = 7,
  UMI_s = NULL,
  recovery_rate = NULL,
  doublet_rate = NULL,
  doublet_factor = NULL,
  planned_read = NULL,
  mapping_efficiency = NULL,
  sideness,
  correction = "BH",
  sig_level = 0.1,
  intermediate_outcome = FALSE
)
}
\arguments{
\item{perturb_type}{Either CRISPRi or CRISPRko; currently can only handle CRISPRi.}

\item{control_cell_vec}{Control cell size for L elements (vector; length L).}

\item{target_cell_mat}{Perturb cell size for L elements (matrix; row: L genomic elements; column: K gRNA libraries). Rownames should be specified for enhancer targets.}

\item{library_size}{Averaged library size (scalar; positive valued).}

\item{relative_expression}{Baseline relative expression level for J genes (named vector; of length J). Names of the vector should be different genes.}

\item{size_parameter}{Size parameter for J genes (vector; length J).}

\item{effect_size_mean}{Mean fold change for each (element, gene) (matrix; row: L genomic elements; column: J genes).}

\item{effect_size_sd}{Sd fold change for each (element, gene) (matrix; row: L genomic elements; column: J genes).}

\item{QC}{A logic value; if TRUE, then do quality control.}

\item{n_nonzero_trt}{QC threshold for treatment cell.}

\item{n_nonzero_ctl}{QC threshold for control cell.}

\item{UMI_s}{Estimate for UMI count per singlet for the underlying type of cell.}

\item{recovery_rate}{Recovery rate for cells surviving the library preparation.}

\item{doublet_rate}{Doublet rate for droplet containing more than one cell.}

\item{doublet_factor}{Ratio of averaged UMI count per doublet to UMI per singlet.}

\item{planned_read}{Planned total sequencing reads.}

\item{mapping_efficiency}{Mapping efficiency for sequenced reads.}

\item{sideness}{Left, right or both.}

\item{correction}{Multiplicity correction; default is BH.}

\item{sig_level}{Significance leve of interest (FDR of interest).}

\item{intermediate_outcome}{A logic value indicating if only mean, sd of test statistics and QC are desired.}
}
\value{
The output format is the same as \code{\link[=compute_power]{compute_power()}}.
}
\description{
Power function for CRISPR screen experimental design.
}
