% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power_function}
\alias{power_function}
\title{Power function for CRISPR screen experimental design}
\usage{
power_function(
  control_cell,
  target_cell_mat,
  UMI_s,
  library_size = NULL,
  expression_level,
  size_parameter_list,
  gRNA_efficiency_mat,
  recovery_rate,
  doublet_rate,
  doublet_factor,
  planned_read,
  mapping_efficiency,
  effect_size,
  sideness,
  correction = "BH",
  sig_level = 0.1,
  return_discovery = TRUE
)
}
\arguments{
\item{control_cell}{Number of cell in the negative control group}

\item{target_cell_mat}{Matrix including number of cell in group (l,k) (row: L genomic elements; column: K gRNA libraries)}

\item{UMI_s}{Estimate for UMI count per singlet for the underlying type of cell}

\item{library_size}{Library size from pilot data or set as a parameter in simulation study}

\item{expression_level}{Expression level for J genes (values between 0 and 1)}

\item{size_parameter_list}{Size parameter list including for all J genes}

\item{gRNA_efficiency_mat}{gRNA efficiency matrix of dimension L by K}

\item{recovery_rate}{Recovery rate for cells surviving the library preparation}

\item{doublet_rate}{Doublet rate for droplet containing more than one cell}

\item{doublet_factor}{atio of averaged UMI count per doublet to UMI per singlet}

\item{planned_read}{Planned total sequencing reads}

\item{mapping_efficiency}{Mapping efficiency for sequenced reads}

\item{effect_size}{Effect size matrix for each element l and gene j (L by J dimension)}

\item{sideness}{Left, right or both}

\item{correction}{Multiplicity correction, example including BH, bonferroni}

\item{sig_level}{False discovery rate level}

\item{return_discovery}{A logic value; if TRUE then returning the discovery set}
}
\value{
Either power list or power list and discovery size
}
\description{
Power function for CRISPR screen experimental design
}
