% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_computation.R
\name{QC_prob}
\alias{QC_prob}
\title{QC probability computation}
\usage{
QC_prob(
  effect_size,
  baseline_expression,
  size_parameter,
  num_control,
  num_trt,
  n_thresh
)
}
\arguments{
\item{effect_size}{Effect size matrix with L row and J column (element, gene)}

\item{baseline_expression}{Baseline gene expression vector of length J}

\item{size_parameter}{Size parameter for gene expression of length J}

\item{num_control}{Number of control cells}

\item{num_trt}{Number of treatment cells (receiving perturbation)}

\item{n_thresh}{Cut off QC size}
}
\value{
QC probability
}
\description{
QC probability computation
}
