% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_help.R
\name{rejection_computation}
\alias{rejection_computation}
\title{Compute the rejection probability.}
\usage{
rejection_computation(mean_list, sd_list, side, cutoff)
}
\arguments{
\item{mean_list}{Asymptotic mean of test statistic}

\item{sd_list}{Asymptotic sd of test statistic}

\item{side}{(Optional) A character string specifying the side of the test, either "left", "right", or "both"; defaults to "both"}
}
\value{
The rejection probablity.
}
\description{
Compute the rejection probability.
}
