% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_help.R
\name{rejection_computation}
\alias{rejection_computation}
\title{Compute the rejection probability}
\usage{
rejection_computation(mean_list, sideness, sig_level)
}
\arguments{
\item{mean_list}{List of mean under local alternative}

\item{sideness}{Sideness of the testing procedure}

\item{sig_level}{Significance level imposed by users}
}
\value{
The rejection probablity
}
\description{
Compute the rejection probability
}
