% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc_help.R
\name{rejection_computation}
\alias{rejection_computation}
\title{Compute the rejection probability of unadjusted score test.}
\usage{
rejection_computation(mean_list, sd_list, side, cutoff)
}
\arguments{
\item{mean_list}{Numeric vector. Mean values of test statistics for each hypothesis.}

\item{sd_list}{Numeric vector. Standard deviation values of test statistics for each hypothesis.}

\item{side}{Character. Test sidedness: "left", "right", or "both".}

\item{cutoff}{(Optional) A numeric value between 0 and 1 to use as the p-value cutoff}
}
\value{
The rejection probablity.
}
\description{
Compute the rejection probability of unadjusted score test.
}
\keyword{internal}
