% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_computation.R
\name{zero_prob}
\alias{zero_prob}
\title{zero probability for each element and gene pair}
\usage{
zero_prob(effect_size, baseline_expression, size_parameter)
}
\arguments{
\item{effect_size}{Effect size matrix with L row and J column (element, gene)}

\item{baseline_expression}{Baseline gene expression vector of length J}

\item{size_parameter}{Size parameter for gene expression of length J}
}
\value{
The zero probability of length L * J
}
\description{
zero probability for each element and gene pair
}
